Frontend wallet actions refactor
