+++ frontend/src/lib/stores/wallet.ts
@@ -7,39 +7,45 @@ export const walletGifts = writable<any[]>([]);
 export const isAdmin = writable<boolean>(false);
 
 if (browser) {
-	const saved = localStorage.getItem('walletAddress');
-	const savedBalance = localStorage.getItem('walletBalance');
-	const savedGifts = localStorage.getItem('walletGifts');
+  // --- Load from localStorage ---
+  const savedAddress = localStorage.getItem('walletAddress');
+  const savedBalance = localStorage.getItem('walletBalance');
+  const savedGifts = localStorage.getItem('walletGifts');
+  const savedAdmin = localStorage.getItem('isAdmin');
 
-	if (saved) walletAddress.set(saved);
-	if (savedBalance) walletBalance.set(savedBalance);
-	if (savedGifts) {
-		try {
-			walletGifts.set(JSON.parse(savedGifts));
-		} catch {
-			walletGifts.set([]);
-		}
-	}
+  if (savedAddress) walletAddress.set(savedAddress);
+  if (savedBalance) walletBalance.set(savedBalance);
+  if (savedGifts) {
+    try {
+      walletGifts.set(JSON.parse(savedGifts));
+    } catch {
+      walletGifts.set([]);
+    }
+  }
+  if (savedAdmin) {
+    isAdmin.set(savedAdmin === 'true');
+  }
 
-	walletAddress.subscribe((addr) => {
-		if (addr) localStorage.setItem('walletAddress', addr);
-		else localStorage.removeItem('walletAddress');
-	});
+  // --- Subscriptions to persist changes ---
+  walletAddress.subscribe((addr) => {
+    if (addr) localStorage.setItem('walletAddress', addr);
+    else localStorage.removeItem('walletAddress');
+  });
 
-	walletBalance.subscribe((bal) => {
-		if (bal) localStorage.setItem('walletBalance', bal);
-		else localStorage.removeItem('walletBalance');
-	});
+  walletBalance.subscribe((bal) => {
+    if (bal) localStorage.setItem('walletBalance', bal);
+    else localStorage.removeItem('walletBalance');
+  });
 
-	walletGifts.subscribe((gifts) => {
-		if (gifts && gifts.length > 0) {
-			localStorage.setItem('walletGifts', JSON.stringify(gifts));
-		} else {
-			localStorage.removeItem('walletGifts');
-		}
-	});
+  walletGifts.subscribe((gifts) => {
+    if (gifts && gifts.length > 0) {
+      localStorage.setItem('walletGifts', JSON.stringify(gifts));
+    } else {
+      localStorage.removeItem('walletGifts');
+    }
+  });
 
-	isAdmin.subscribe((val) => {
-    	localStorage.setItem('isAdmin', val ? 'true' : 'false');
-  	});
+  isAdmin.subscribe((val) => {
+    localStorage.setItem('isAdmin', val ? 'true' : 'false');
+  });
 }
