+++ frontend/src/routes/createListing/[id]/+page.svelte
@@ -3,7 +3,7 @@
   import { goto } from "$app/navigation";
   import { walletAddress } from "$lib/stores/wallet";
   import { get } from "svelte/store";
-  import { getWalletFromMnemonic, signedFetch } from "$lib/walletActions";
+  import { mnemonicMatchesLoggedInWallet, signedFetch } from "$lib/walletActions";
   import MnemonicInput from "$lib/MnemonicInput.svelte";
 
   import { page } from "$app/stores";
@@ -92,8 +92,7 @@
     }
 
     try {
-      const wallet = getWalletFromMnemonic(mnemonic);
-      if (wallet.address.toLowerCase() !== address.toLowerCase()) {
+      if (!mnemonicMatchesLoggedInWallet(mnemonic)) {
         error = "Mnemonic does not match the logged-in wallet";
         return;
       }
@@ -125,7 +124,7 @@
           headers: { "Content-Type": "application/json" },
           body: JSON.stringify(listing),
         },
-        wallet,
+        mnemonic,
       );
 
       if (!res.ok) throw new Error("Listing failed");
